requirements.txt
functions-framework==3.*
Flask>=2.2.2
tweepy>=4.10.1
pymediawiki>=0.7.2
python-dotenv>=0.21.0
replicate>=0.4.0